Fixed (hopefully) the double rounding issues

+++ dog/dog/out.cpp
@@ -6,7 +6,7 @@ out::out(std::string toOutput)
 	std::cout << toOutput << std::endl;
 }
 
-out::out(int toOutput)
+out::out(double toOutput)
 {
 	std::cout << toOutput << std::endl;
 }
